Fine_Tuning_LLama3_To_NPC/requirements.txt
absl-py==2.2.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
Appium-Flutter-Finder==0.7.0
Appium-Python-Client==4.0.1
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
asyncpraw==7.8.1
asyncprawcore==2.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
behave==1.2.6
boto3==1.35.19
botocore==1.35.19
cachetools==5.4.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
discord==2.3.2
discord.py==2.4.0
distlib==0.3.8
distro==1.9.0
easy-pil==0.4.0
executing==2.2.0
Faker==37.3.0
filelock==3.15.4
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.4.1
fsspec==2025.3.2
gast==0.6.0
gherkin-official==29.0.0
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.63.2
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.3.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
openai-cli==0.0.3
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parse==1.20.2
parse_type==0.6.4
parso==0.8.4
pdfkit==1.0.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
playwright==1.50.0
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.24.0
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.3
pydantic_core==2.18.4
pyee==12.1.1
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.3.2
pytest-bdd==8.1.0
pytest-html-reporter==0.2.9
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.3.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9
s3transfer==0.10.2
scikit-learn==1.6.1
scipy==1.15.3
selenium==4.23.1
setuptools==78.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-macos==2.16.2
termcolor==3.0.1
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.66.4
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.3
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
